datapath: move fake.GNeighConfig to pkg/datapath/gneigh/fake.Config
